fix: invalid value serialization when input property is not provided
When an input property is not provided its value won't be set to null,
it will be set to undefined. This would cause the ValidateDirectiveVisitor
to try to serialize the value or in case of more complex types, try to
validate it as well.
In order to avoid these cases, undefined must also be tested.

Closes #10