feat: move validationErrors to GraphqlResolverInfo
This commit aims to do not pollute the GraphQL argument field
on the resolver. By doing so, one can omit the validationErrors
argument from the schema, since this field should only be
used internally to know if a validation has failed.
In order to achieve this feature, this commit will
inject the validationErrors field directly
in the GraphQLResolverInfo argument.