feat: add policy argument to ValdiateDirectiveVisitor
This new arg will not hide an validation errors on
arguments and inputs in case they are nullable.
In such cases an valÃidation error will be thrown.
In order to do not break the API, this patch
will leave this new argument with a default value
which matches the previous behavior.