fix: properly throw on validations erros when policy is THROW
Prior to this patch all directives were failing to throw the
validation error if the policy was set to THROW and the
value itself was deep inside inputs. This was happening,
because the policy was being set on the container itself
on not on the field. Also there was a bug that could also
cause the ValidateDirectiveVisitor to do not throw errors
if you have two directives on the same input with
different policies.
In order to prevent such scenarios, this comit attaches
the policy in the field itself and marks the thrown error
with a flag telling the parent types that this error
can not be silecend.