fix: do not change validated values in-place
Changing values in place may cause problems for some directives,
thus breaking the user's code. For example image the following
schema:

type Query {
   sellBook(id: ID! @foreignNodeId(typename: "Book")): SellData
   getBook(id: ID! @foreignNodeId(typename: "Book")): Book
}

and a query

query getBooks($id: ID!) {
   sellBook(id: $id) {
     id
   }
   getBook(id: $id) {
     id
   }
}

Prior to this patch, an error would be thrown to the user,
since the @foreignNodeId directive would first convert
the 'id' for the sellBook mutation and then convert
it *again* for the getBook mutation, this would
impose a problem since the convertions are done
in-place and the id variable was already converted!
For example:

* id is set to 'abcde' by the user before executing the mutation
* @foreignNodeId directive on sellBook mutation converts 'abcde' to 'edcba' (this is
in-place so no the id value will be edcba on all it usages)
* @foreignNodeId directive on getBook mutation tries to convert 'edcba' (this is wrong, the
id was already converted)
* BOOM.

In order to avoid such problems, this patch changes the in-place convertion
to threat objects as immutables. Thus creating new objects everytime
a convertion is made.