fix: allow permission validation to be skipped on null
This commit adds the support for @hasPermissions directive
to be skip the permission validation if the value of an input field/input
or argument definition value is null or undefined.
This type of check if useful for when @hasPermissions is
used on optinal fields in a input, thus leaving room for
one to continue to use the input even though he/she
may not have the permission to use that field.